Fix rate limiting, AFL blowout match, stoppage time parsing
- Cache /api/live-games server-side (15s TTL) to stop tripling API calls
- Fix AFL incorrectly matching "football" blowout tier (use startswith)
- Fix soccer stoppage time with decimal minutes (use float not int)
- Set max_bet_pct=15%, NBA 5:00/11pt via config API

+++ espn.py
@@ -307,7 +307,7 @@ async def get_scoreboard(sport_path: str) -> list[GameState]:
             # Handle soccer stoppage time: "45'+4'" → 45+4 = 49 minutes
             if "+" in clean and "soccer" in sport_path:
                 parts = clean.replace("'", "").split("+")
-                total_mins = sum(int(p.strip()) for p in parts if p.strip())
+                total_mins = sum(float(p.strip()) for p in parts if p.strip())
                 clock_seconds = total_mins * 60
             else:
                 # Strip trailing apostrophes
